feat: altを表示する仕組みをRecyclerViewのロジック内から切り出した
