+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/view/NoteCardActionListenerAdapter.kt
@@ -4,6 +4,7 @@ import net.pantasystem.milktea.model.channel.Channel
 import net.pantasystem.milktea.model.note.Note
 import net.pantasystem.milktea.model.note.poll.Poll
 import net.pantasystem.milktea.model.user.User
+import net.pantasystem.milktea.note.media.viewmodel.PreviewAbleFile
 import net.pantasystem.milktea.note.reaction.ReactionCountAction
 import net.pantasystem.milktea.note.viewmodel.PlaneNoteViewData
 
@@ -64,6 +65,9 @@ class NoteCardActionListenerAdapter(
         onAction(NoteCardAction.OnChannelButtonClicked(channelId))
     }
 
+    fun onMediaPreviewLongClicked(previewAbleFile: PreviewAbleFile?) {
+        onAction(NoteCardAction.OnMediaPreviewLongClicked(previewAbleFile))
+    }
 }
 
 
@@ -82,4 +86,6 @@ sealed interface NoteCardAction {
     data class OnUserClicked(val user: User) : NoteCardAction
     data class OnFavoriteButtonClicked(val note: Note) : NoteCardAction
     data class OnChannelButtonClicked(val channelId: Channel.Id) : NoteCardAction
+
+    data class OnMediaPreviewLongClicked(val previewAbleFile: PreviewAbleFile?) : NoteCardAction
 }
