+++ modules/features/note/src/main/res/layout/item_simple_note.xml
@@ -228,6 +228,7 @@
                         android:id="@+id/manyFilePreviewListView"
                         mediaViewData="@{note.media}"
                         previewAbleList="@{note.media.files}"
+                        noteCardActionListenerAdapter="@{noteCardActionListener}"
                         android:layout_marginTop="@dimen/note_media_preview_top_margin_size"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
@@ -365,6 +366,7 @@
                             android:layout_below="@id/subNoteText"
                             mediaViewData="@{note.subNoteMedia}"
                             previewAbleList="@{note.subNoteMedia.files}"
+                            noteCardActionListenerAdapter="@{noteCardActionListener}"
                             tools:layout_height="20dp"
                             tools:visibility="gone"
                             rect="@{20f}"
